Refactor sycl runtime build
